olivbackend/pyproject.toml
[project]
name = "olivbackend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "duckduckgo-search>=8.1.1",
    "google-genai>=1.67.0",
    "python-dotenv>=1.2.2",
]
